Previously was pushing all of the Local MD constructed potentials to the end of the list of potentials. However this creates concurrency issues due to the syncing that takes place. Instead moving the potentials without syncs to the starts allows us to hide more latency.

* Improves performance by ~4% in solvent, ~2% in complex.

# Benchmarks

A4000 GPU
Cuda Arch 8.6

## Master
```
hif2a-rbfe-local: N=8840 speed: 1252.14ns/day dt: 2.5fs (ran 100000 steps in 17.26s)
solvent-rbfe-local: N=6317 speed: 1518.21ns/day dt: 2.5fs (ran 100000 steps in 14.23s)
```

## PR
```
hif2a-rbfe-local: N=8840 speed: 1272.77ns/day dt: 2.5fs (ran 100000 steps in 16.98s)  # 2% speed up
solvent-rbfe-local: N=6317 speed: 1568.31ns/day dt: 2.5fs (ran 100000 steps in 13.78s) # 4% speed up
```