@badisa 2024-09-03
Probably going to be a net negative with https://github.com/proteneer/timemachine/pull/1371 since we want the nonbonded kernels to be upfront since we can hide the faster kernels in the shadow of the Nonbonded kernels.